+++ R/class-auth.R
@@ -2,6 +2,8 @@
 #' 
 #' Use firebase to manage authentications.
 #' 
+#' @return An object of class \code{FirebaseAuth}.
+#' 
 #' @importFrom cli cli_rule cli_alert_danger cli_alert_info cli_alert_warning cli_alert_warning
 #' @importFrom jsonlite fromJSON
 #' @importFrom jose jwt_decode_sig
@@ -40,12 +42,19 @@ FirebaseAuth <- R6::R6Class(
     },
 #' @details Print the class
     print = function(){
-      rule("Firebase Authentication")
-      signin <- "No user is signed in"
-      if(private$.user_signed_in$signed_in)
-        signin <- "A user is signed in"
-      
-      cli_alert_info(signin, "\n")
+      cli_rule("Firebase Authentication")
+      user <- self$get_signed_in()
+      if (!isTRUE(user$success)) {
+        cli_alert_info("No user is signed in")
+      } else {
+        cli_alert_info("A user is signed in")
+        cli::cli_dl(Filter(Negate(is.null), c(
+          uid = user$response$uid,
+          email = user$response$email,
+          emailVerified = user$response$emailVerified,
+          displayName = user$response$displayName
+        )))
+      }
     },
 #' @details Signs out user
 #' @return self